+++ activesupport/lib/active_support/core_ext/symbol.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
+require "active_support/core_ext/symbol/inflections"
 require "active_support/core_ext/symbol/starts_ends_with"